requirements.txt
django==5.0.6
djangorestframework==3.15.1
djoser==2.2.2
django-cors-headers==4.3.1
django-environ==0.11.2
django-storages==1.14.2
drf-yasg[validation]==1.21.7
psycopg2==2.9.9
psycopg2-binary==2.9.9
pillow==10.3.0
dj-static==0.0.6
static==1.1.1
static3==0.7.0
django-filter==24.2
celery==5.4.0
django-celery-beat==2.6.0
django-celery-results==2.5.1
redis==5.0.4
hiredis==2.3.2
django-cloudinary-storage==0.3.0
cloudinary==1.40.0
